Use 8-bits textures for text atlas

+++ swGraphicAPI/Assets/TextAsset/Loader/FontLoader.h
@@ -42,7 +42,7 @@ public:
 public:
     /**@brief Renders Texture Atlas. Modifies FontLayout by adding texture mapping.*/
     Nullable< TexturePtr >          RenderAtlas         ( const FreeTypeLibrary& freeType, const AssetPath& filePath, FontLayout& initData, RMLoaderAPI factor );
-    Image< u32 >                    RenderAtlasToBuffer ( const FreeTypeLibrary& freeType, FontLayout& initData, uint32 width, uint32 height );
+    Image< u8 >                     RenderAtlasToBuffer ( const FreeTypeLibrary& freeType, FontLayout& initData, uint32 width, uint32 height );
 };
